requirements.txt
google-genai>=1.0.0
feedparser>=6.0.12
playwright>=1.58.0
pydantic>=2.0.0
